Extracted zjsunit/render.js
